Changed code to work with bytes. Small bug fixes for maze instantiation

+++ arduino/micromau5/coord.h
@@ -1,5 +1,5 @@
 struct coord
 {
-  int x;
-  int y;
+  byte x;
+  byte y;
 };
